backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
Pillow==10.0.0
PyPDF2==3.0.0
requests==2.31.0
aiofiles==23.2.1
websockets==12.0
python-dateutil==2.8.2
aiohttp==3.9.0
sqlalchemy==2.0.23
alembic==1.13.0
